feat(authentification): Header generic
